revert text handling defaults
when the user passes a URL, how should it be handled?  the "simplest" default behavior
would be to assume that URL-like text refers to a remote object.  however, what if the
user instead wants to pass in text that happens to be a URL?  this could lead to
strange or unexpected behaviors.  instead, we'll assume that *all* text should be
interpreted by the `wrangle` function as text.  however, if the user wants to load in
a local or remote file, we also provide the `datawrangler.io.load` function for handling
local or remote files of a wide variety of types, and returning objects that can be
wrangled.

i'm not sure this is an optimal solution, but at least it avoids potentially dangerous or
unexpected behavior...